+++ src/misc/subpixel.h
@@ -1,3 +1,3 @@
 
-extern void est_subpixel_shift(unsigned int N, float shifts[N], const long dims[N], unsigned int flags, const complex float* in1, const complex float* in2);
+extern void est_subpixel_shift(int N, float shifts[N], const long dims[N], long flags, const complex float* in1, const complex float* in2);
